requirements.txt
openai==0.27.0
google-api-python-client==2.92.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-dateutil==2.8.2
